requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-lock.txt --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
backports-asyncio-runner==1.2.0
    # via pytest-asyncio
blinker==1.9.0
    # via flask
cachetools==6.2.6
    # via pyiceberg
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   pyiceberg
cryptography==46.0.5
    # via
    #   google-auth
    #   pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
exceptiongroup==1.3.1
    # via
    #   anyio
    #   pytest
flask==3.1.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   pytest-flask
flask-cors==6.0.2
    # via -r requirements.in
fsspec==2026.2.0
    # via pyiceberg
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.30.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.190.0
    # via google-generativeai
google-auth==2.48.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==25.1.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   -r requirements.in
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via pyiceberg
multidict==6.7.1
    # via yarl
numpy==2.2.6
    # via
    #   -r requirements.in
    #   shapely
packaging==26.0
    # via
    #   deprecation
    #   gunicorn
    #   pytest
pluggy==1.6.0
    # via pytest
postgrest==2.28.0
    # via supabase
propcache==0.4.1
    # via yarl
proto-plus==1.27.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   google-generativeai
    #   postgrest
    #   pydantic-settings
    #   pyiceberg
    #   realtime
    #   sqlmodel
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.in
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyiceberg==0.11.0
    # via storage3
pyjwt==2.11.0
    # via supabase-auth
pymupdf==1.27.1
    # via -r requirements.in
pyparsing==3.3.2
    # via
    #   httplib2
    #   pyiceberg
pyroaring==1.0.3
    # via pyiceberg
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-flask
pytest-asyncio==1.3.0
    # via -r requirements.in
pytest-flask==1.3.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via strictyaml
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
realtime==2.28.0
    # via supabase
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   pyiceberg
rich==14.3.3
    # via pyiceberg
rsa==4.9.1
    # via google-auth
shapely==2.1.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.47
    # via sqlmodel
sqlmodel==0.0.37
    # via -r requirements.in
storage3==2.28.0
    # via supabase
strenum==0.4.15
    # via
    #   postgrest
    #   supabase-functions
strictyaml==1.7.3
    # via pyiceberg
supabase==2.28.0
    # via -r requirements.in
supabase-auth==2.28.0
    # via supabase
supabase-functions==2.28.0
    # via supabase
tenacity==9.1.4
    # via pyiceberg
tomli==2.4.0
    # via pytest
tqdm==4.67.3
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   anyio
    #   cryptography
    #   exceptiongroup
    #   google-generativeai
    #   grpcio
    #   multidict
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   realtime
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
websockets==15.0.1
    # via realtime
werkzeug==3.1.6
    # via
    #   flask
    #   flask-cors
    #   pytest-flask
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-functions
zstandard==0.25.0
    # via pyiceberg
